Feedback:
Comment Received:
Nice thought. However, I think the data gotten from the API is unique to each device. So the possibility of it being updated from another is slim. Using await here also slows down the bookshelf change. It means that the BooksAPI has to be updated first before the shelf change is reflected on the browser. While this works, it causes the change to happen slowly and that may not create a good user experience. You may want to change the state first before making a request to the API.

Update:
Alright. Updating the state first and then calling the POST method to update the backend later.