Feedback: Search does not work as expected
Comment Received:
the page does not meet the (a) requirement. This requirement means that as the user types, books that match the query are displayed. However, when I type as shown in the picture below, nothing is displayed.This happens because, in your code, you restricted the search to work for only keywords. However, if the user does not know or type a keyword, nothing is displayed. Please update this to allow the search results to display for any input.

Update:
Incorporated. Comments added in the code to explain the logic.